requirements.txt
streamlit==1.29.0
anthropic==0.3.3
python-dotenv==1.0.1
Pillow==10.0.0
validators==0.21.0
